prov/opx: DAOS: CART server-client tests fail after restart of client
Currently, the DAOS server and client test applications can successfully exchange
any quantity of messages with each other.  When the client successfully completes
all operations and terminates and the server continues to run, if the client is
restarted, all packets sent from the client fail due to time-out.

Unlike MPI, DAOS uses the server-client application communication model.  The
current implementation of the Reliability Protocol (RP) of OPX inhibits proper
support for this application model.
For each endpoint (EP) communication interaction, the RP maintains addressing and
PSN related data on each EP.  The MPI application model assumes that all EPs of
the application will start and terminate together.  For DAOS this is not so,
client/server EPs (especially the client) randomly start, terminate, and re-start
often.

To provide proper support for DAOS in OPX, enhancements are required to enable the
resynching/resetting of the RP data maintained on each EP.

GENERAL ENHANCEMENT DESCRIPTION
-------------------------------
• New reliability protocol UD opcodes RESYNCH & ACK RESYNCH
   o Supported via in-band (HFI)
       - [INIT] Extended PIO send model initialization functionality
       - [SEND] Extended the fi_opx_hfi1_tx_reliability_inject () functionality
       - [RECV] Extended fi_opx_hfi1_handle_ud_packet() functionality
       - New UD opcodes operate independent from other RP UD opcodes.  They do utilize
         the transport mechanism (extensions previously mentioned).  They do not
         utilize the RP framework/infrastructure
   o Supported via SHM
       - SHM interfaces extended to support the new UD opcodes
• Server EP role:
   o Server EP the only one to receive and process the RESYNCH request
   o Server EP the only one to send the ACK RESYNCH response
• Client EP role:
   o Client EP the only one to send RESYNCH request
   o Client EP the only one to receive ACK RESYNCH response

RESYNCH ENHANCEMENT WORK-FLOW DESCRIPTION
-----------------------------------------
1. Server EP comes up and waits for connection request from local/remote Client EP.
2. Client EP comes up and attempts to establishes connection to local/remote Server EP
3. First, the Client EP sends RESYNCH request to Server EP and waits for ACK RESYNCH
   response
       a. Until the Client EP successfully receives the ACK RESYNCH response, it will
          not send any other control/data packets to the Server EP
4. Server EP receives RESYNCH request and resets all the in-band and SHM reliability
   protocol related data associated with the Client that is maintained locally by the Server EP
5. Server EP sends ACK RESYNCH response after successfully resetting all data
6. Client EP receives ACK RESYNCH response and then starts normal transmission of all other
   control/data packets to the Server EP

Signed-off-by: Charles King <charles.king@cornelisnetworks.com>