prov/shm: Implement shared completion queues
On cq_open, check the FI_PEER_IMPORT, if set, set all internal cq operation
to be enosys, with the exception to the read callback.

The read callback is overloaded to operate as a progress callback
function. Invoking the read callback will progress the enpoints linked to
this CQ.

Keep track of the fid_peer_cq structure passed in.

Update the shm cq to be it's own structure, which embeds the util_cq, as
well as a pointer to the fid_peer_cq and a callback used for writing into
the CQ, whether internal or imported.

If the FI_PEER_IMPORT flag is set, then set the callbacks in smr_cq structure
which handle writing to the peer_cq, otherwise set it to the ones which
write to the util_cq. These callbacks are called at the bottom of the stack
which is triggered via the tx_comp()/rx_comp() call flow.

The order of operations is as follows:
. open the CQ, if it is imported, then set internal callbacks to enosys
  with the exception of the read callback as explained above.
. create a new smr_cq structure and store a pointer to the fid_peer_cq
  passed in the open call.
. When the CQ is bound to an ep, insert the endpoint on the CQ ep_list
. In the call flow fi_read_cq->smr_peer_cq_progress->ofi_cq_progress
   - This will walk the EPs on the cq list and call ep->progress() for
     each
   - ep->progress() will kick off handling any queued up work and
     calls the ep->rx_comp() or ep->tx_comp() for the completed requests.
   - These end up calling the peer specific completion callbacks registered
     earlier.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>