prov/efa: remove struct efa_ep_domain
The struct efa_ep_domain is neither an ep nor a domain, but
a collection of 3 piece of information about a domain to be
created: ep_type, name_suffix and capability.

This patch removes the struct efa_ep_domain. As for the
3 memebers of efa_ep_domain:

ep_type and capability is directly accessed via variables.

This patch introduced a function efa_domain_name_suffix() to
return the name suffix of different domain type.

Signed-off-by: Wei Zhang <wzam@amazon.com>