core: Allow associating a user identifier with an AV address
This feature targets the peer provider use case, but may be useful
for other applications over specific providers.  Allow the user to
assign an identifier (e.g. a pointer to some structure) to
an inserted AV address.  Instead of libfabric reporting the AV's
fi_addr_t value as the src address for a completion, the provider
will return the user's identifer instead.  This can avoid a reverse
lookup by the user.

When combined with the peer provider, this optimizes the receice
and completion handling by avoiding a lookup in the owner provider,
since the peer (i.e. shm) has already performed a lookup, and it's
just as easy for the peer to return the user's id as the fi_addr_t
associated with the AV.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>