prov/efa: introduce function efa_prov_info_alloc_rxr()
A prov_info contains the attributes of a provider,
and is needed to valiadate the info passed by
user.

Currently, the prov_info of the RDM endpoing of EFA is a global
variable named rxr_info.

When it is initialized, rxr_info does not have all the attributes,
because some attributes are from EFA device.

Currently, these device specific attributes in rxr_info was set
when user call fi_getinfo, which call rxr_info_to_rxr().
rxr_info_to_rxr() update rxr_info which the device information.

The update of a global variable during runtime is undesirable.
To fix it, this patch makes the following changes:

It introduced a function efa_prov_info_alloc_for_rxr(), which
will allocate a prov_info that matches the attributes of EFA
RDM endpoint.

This function was called during provider initialization, and
the allocated prov_info_rxr was stored in prov_util.

The code in rxr_info_to_rxr() that update rxr_info was moved
into efa_prov_info_alloc_for_rxr().

The global variable rxr_info was removed.

The attributes in file rxr_attr.c was merged into
efa_prov_info_alloc_for_rxr(), the rest was move to rxr_prov.c

Signed-off-by: Wei Zhang <wzam@amazon.com>