prov/linkx: Implement fi_getinfo()
Implement fi_getinfo(). It looks for all providers which have the FI_LINK
capability. It then pairs these with the SHM provider given that it also
support FI_LINK.

It represents each entry in the info list as follows:

  provider: shm+<prov>;ofi_linkx
      fabric: ofi_lnx_fabric_3
      domain: shm+<prov>;ofi_lnx_domain
      version: 115.0
      type: FI_EP_RDM
      protocol: FI_PROTO_LINKX

This allows existing applications to make use of the LINKx provider
without having to make any changes.

This method will not support linking more than one provider with the shm
provider dynamically. There doesn't seem to be a good way to represent all
the different of provider domains which can be linked
together.

For example if you have 4 tcp interfaces and you want to link 2
of these interfaces along with the shm provider, how can the application
specify the two interfaces it needs to link?

With the current method fi_getinfo() provides, we'll need to present all
the permutation of possible linking, which is not feasible.

To handle this case, the fi_link() API is added to allow the application
to specify the exact set of provider endpoints to link together.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>