core/pollfds: Use a compact pollfd array
The mdtest benchmark is showing significant performance regressions
as a result of commit 7ec82992.  That commit replaced a compact
pollfd array, where all entries are valid, with a sparsely filled
array.  This removed a linear search through the array in order
to update events.

Rather than revert the above commit, we separate the pollfd
context array and pollfd array.  The fd may still be used to
directly index into the context array.  The context structure is
updated to store the corresponding index where the fd is located
in the pollfd array.  This avoids the linear search while
maintaining a compact array.

As a result of these changes, the windows and unix implementations
can align to share more code.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>