log: Remove dependency to config.h from fi_log.h
Patch a33a3e577("log: Add the provider header files to the
installation") added the header file fi_log.h to the installation.
Problem is that fi_log.h depends on contrib.h, but the application
cannot find it because it's not installed.

The fix implemented in the patch is to remove the inclusion of fi_log.h
and to use #ifdef instead of #if.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>