prov/util: fix IP util av default addrlen
When creating a IP util_av, if we do not yet know whether we are using IPv4 or IPV6,
the address format will be set to FI_SOCKADDR. In order to allocated enough space
in each buffer entry, the ofi_ip_av_create_flags call defaults to the larger
address length. However, when actually inserting addresses, we can determine the
address format by looking at the address being inserted.

This patch adds an assert to make sure the entry size is large enough to accomodate
the inserted address and sets the av->addrlen to be the correct address length.

On the rxm path, when initializing the util av, rxm sets the addrlen from the results
of ofi_sizeof_addr_format which currently does not support FI_SOCKADDR. As a result,
rxm was passing in an addrlen of 0, preventing the proper amount of space in the buffer
pool from getting allocated. This patch also adds FI_SOCKADDR to that size lookup to
adequately allocate memory for the addresses.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>