prov/efa: argument name change for rxr_pkt_init_data_from_tx_entry()
This patch renamed the argument "hdr_size" to "pkt_data_offset",
because "pkt_data_offset" precisely define the nature of
the argument. For upcoming runting REQ packets,
hdr_size does not equal to the data offset in the packet.

Meanwhile, it renamed "data_offset" to "tx_data_offset" to be
more precise

Signed-off-by: Wei Zhang <wzam@amazon.com>