prov/verbs: set and check address format correctly
When building the list of fi_infos, the address formats should be
set since FI_EP_MSG only supports FI_SOCKADDR formats and UD
only supports FI_ADDR_IB_UD format. Set these early on and then
add a check when comparing the fi_infos and hints to make sure
the address format for that fi_info correctly matches the hints
format if it was specified.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>