prov/efa: simplify rxr_read_alloc_entry() using rxr_op_entry
This patch simplifies the function rxr_read_alloc_entry() by
changing the argument x_entry's type from "void *" to
"struct rxr_op_entry *".

The type was "void *" because the input can be a tx_entry
or an rx_entry. After the type "rxr_tx_entry" and "rxr_rx_entry"
is merged into "rxr_op_entry". The argument can be of
type "rxr_op_entry", and the implementation can be simplified.

Signed-off-by: Wei Zhang <wzam@amazon.com>