prov/opx: Fix hang in intranode fence operation
When attempting an intranode fence operation and
the TX FIFO is full, we poll in order to free up some
slots. If the next received packet is a fence operation,
we then loop forever trying to get another TX FIFO slot.

This change makes the receiving fence operation deferrable
instead of looping forever, so we can move on to the next
received packet and make some progress.

Signed-off-by: Ben Lynam <ben.lynam23@gmail.com>