fabtests/av_test: add support for -e <ep_type>
The av_test supports both rdm and dgram EP types, defaulting to RDM but
falling back to DGRAM if it does not get selected. This is redundant
because DGRAM providers will probably be able to layer under rxd which
will let the RDM code work.

This change lets the user select which one to run with the -e argument
which allows the test to run with DGRAM only providers that would
otherwise use rxd on top.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>