prov/cxi: Handle C_EVENT_SEND events
Even though send events are disabled, hardware can generate a send event
in addition to the normal put/get/amo event when an error occurs. Prior
to this, a sent event would cause an abort to be tripped. This is
incorrect. The correct action is to just drop the send event and wait
for the normal event.

NETCASSINI-3774

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>