prov/efa: rename protover in rxr_req_info to extra_info_id
The name "protover" came from a deprecated interpretation of the RDM
protocol. In that interpretation, each version of protocol
will have a 64 bits capability flag, an extra feature
will be assigned a bit in the capability flag.

REQ packet types can be from different version. "protover" indicates
the protocol version of the REQ packet type belongs to, and is used
to find its extra feature flag.

This interpretation has been deprecated. The current interpretation
is that protocol version will not change (always 4). But, version
4 can have multiple 64 bits flags, stored in an array named
"extra_info".

This patch rename the field "protover" in rxr_req_info to
"extra_info_id", and use it to find flag extra_info.

A detailed explaination about this change can be found in
Section 2.1 of the EFA rdm protocol v4 document:

https://github.com/ofiwg/libfabric/blob/main/prov/efa/docs/efa_rdm_protocol_v4.md

Signed-off-by: Wei Zhang <wzam@amazon.com>