core/pollfds: Fix windows ofi_pollfds_do_del()
The implementation of ofi_pollfds_do_del() does not work with
the windows version of pollfds when removing an entry from the
middle of the fds.  The issue is that the fds[] and ctx[]
arrays are kept in sync, unlike linux.  So, when swapping an
fds[] entry we also need to swap the ctx[] entries.

To fix, we update the windows implementation to allow the
fds[] and ctx[] arrays to be independent from each other.  When
allocating a ctx[], we search the array for an unused entry.
The get_ctx[] call is also updated to search all ctx[] entries,
since the matching ctx may be beyond the nfds count.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>