include/rdma: modify peer_srx API
Add the following fields to the shared receive entry:
- srx pointer to get back to the srx that owns it/is processing a message for it
- owner_context for the owner/allocator to manage resources
- user_context for the user (when the entry represents an incoming message,
not a posted receive)
- flags for a peer to know whether to write completions or not
- change iov[] to iov* - the allocator should be responsible for allocating the
appropriate iov array and can set this pointer to point to where to find the array

Separate the get_msg and get_tagged functions into the following:
- get_msg/get_tagged, queue_msg/queue_tag, free_entry
This allows the peer to allocate and set the rx_entry user_context when an unexpected
message arrives so it has the needed information when the owner calls start after
getting an appropriate posted receive.
Since the owner is always responsible for allocating the rx_entry, free_entry allows
the peer to tell the owner when it is done with the owner's rx_entry

The peer ops remain the same other than removing the srx entry from the parameters
since it is now in the entry.

The flow should be as follows:
On fi_recv:
- owner allocates rx_entry and checks unexpected queue, calling start_msg if a match
is found

On incoming message:
- peer calls get_msg which will always return an rx_entry, whether a match was found
or not - if a match was found, return FI_SUCCESS, if no match, allocate entry and
return FI_ENOENT
- if no match was found, peer allocates any necessary resources for processing the
unexpected message later and saves user_context appropriately. if a match was found,
peer can begin processing the message and complete according the the rx_entry flags

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>