fabtests/rdm_stress: Add debug option
The rdm_stress client will fork multiple processes to
communicate with the server.  If an error occurs in one
of those processes, debugging the problem is difficult.
Add a --debug option which will convert all asserts into
while loops.  This will cause the process to spin until
a debugger can be attached.

Relocated all long opt related structs and functions
together as part of this change.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>