prov/opx: Fix bad pointer usage bug
The function fi_opx_ep_tx_send_rzv was putting a pointer
to a stack-frame object in the network packet headers.
This is not an acceptable use of memory and could cause
random corruption in user-space.

Fix the function to handle the case of a null origin byte
counter in the correct way.

Signed-off-by: Tim Thompson <tim.thompson@cornelisnetworks.com>