prov/efa: send EOR earlier for RUNTREAD
Currently, for RUNTREAD, receiver sent the EOR packet when all the
data is copied to receiving buffer.

This patch let the receiver to send EOR when the RDMA read is
finished to improve efficiency.

With this change, runtread message protocol does not support
delivery complete any more. Therefore, this patch also changed
the protocol selection logic that runt read is NOT used if
delivery complete has been requested.

Signed-off-by: Wei Zhang <wzam@amazon.com>