core/indexer: Replace ARRAY_SIZE with MAX_CHUNKS
The indexer manages an array of an array of items.  The ARRAY_SIZE
define is actually the size of the first array, which are the
chunks of memory to allocate.  Rename so that the code better
indicates which type of array we're dealing with.

This change highlights an incorrect use of ARRAY_SIZE, which
is left unchanged for now.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>