prov/shm: move shm signal handlers initialization to EP
Currently, any use of libfabric will cause shm to add its
signal handlers, even if shm is not being used.

This moves the signal handler initialization to fi_endpoint so
we only add them if they are needed.

The old_action free is kept in the shm FINI call since we need
to wait for all EPs to be closed.

This also fixes a typo in the function smr_reg_sig_handler and
fixes some function and variable definition inconsistencies.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>