prov/linkx: Support memory registration
Add support for fi_mr_regattrs() API in LINKx provider.

Upon handling fi_mr_regattrs() call, LINKx will check the domain's
mr_mode. If there are no flags set or if FI_MR_ENDPOINT is set, then don't
register the memory. Otherwise call the core's memory registration. This
is to adhere with the behavior outlined here:

https://ofiwg.github.io/libfabric/v1.15.0/man/fi_mr.3.html

LINKx needs to determine which domain to register the memory against.
Since it handles multiple domains, the only way it can find out the
domain is by being provided a destination or source address. It can then
use the address to predetermine the local domain it'll use for the data
operation and register the memory against it.

Therefore, fi_addr_t field has been added to fi_mr_attr structure to
allow LINKx to perform the above logic.

If the address is not specified, then LINKx will need to register the
memory with all the domains it has, because it doesn't know which domain
will endup handling the data operation.

TODO: I'm unsure if registering the same piece of memory with many domains
is going to cause a problem.

The fid_mr obtained from the fi_mr_regattr() call to the core provider
as well as the local endpoint which will be used for the pending data
operation and the core provider peer address will be stored in the LINKx
memory descriptor, which will be returned in the fid_mr structure in the
fi_mr_regattr() result.

The application will then send this information to the LINKx provider in
the data operation. The LINKx provider can then use this information to
determine the local endpoint, peer address and fid_mr to forward to the
core provider.

In the absence of the descriptor, then LINKx will go through the standard
procedure to determine the local endpoint and the peer address.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>