fabtests/rdm_stress: Fix timeout error handling hello message
If we fail to start a hello rpc, for example the peer has exited
prior to receiving a response, we wll remove the peer's address
from the AV and retry up to the timeout window.  After we time
out, we'll complete the hello rpc as a failure.  Since the rpc
failed, we'll try to remove the peer's address from the AV.
This causes a segfault in libfabric because the address is
invalid.

Handle this case by setting the client_id to -1 anytime we're
still in need of an address.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>