prov/shm: Fix check for FI_MR_HMEM mr_mode
In smr_getinfo, we perform a post-getinfo check that the FI_MR_HMEM
mr_mode bit is set correctly if FI_HMEM has been requested.  However,
there may be valid fi_info structures to return.  By failing
the smr_getinfo call, we leak those structures and fail to return
them as valid.

To fix this we add a domain_attr to pair with the FI_HMEM info
structures, so that util_getinfo can handle the filtering.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>