Fetch Slingshot product version from branch
Fetch the Slingshot product version from a branch, so that release
branch builds get the right product version.

NB: this commit should not be merged into the release/slingshot-1.7
branch, as this will cause the RPM version to go backwards and cause
packaging issues. Only merge this commit into Slingshot 2.0.0 and
beyond.

Fixes: NECASSINI-3841

Signed-off-by: Richard Halkyard <richard.halkyard@hpe.com>