Add zbcoll reduce function.
NETCASSINI-3863

Add a caller-accessible cxip_zbcoll_reduce() function to allow access to
the same bitwise-AND reduction used in getgroup. This allows an error
flag to be contributed from all endpoints, and distributed back to all
endpoints.

Extend zbcoll testing.
Disable coll testing (to pass regression).

Fix bug in zbsend_up() where the result of zbpack() is never used.
Appears to have functioned in tests because of the limited requirements
of the bitwise-AND for getgroup, and incomplete test coverage.
Generalized bitwise-AND with fully randomized testing uncovered this.

Restructure code to clarify: the bitwise-AND is no longer done in
zbsend_up(), but instead on receipt of the data value in
cxip_zbcoll_recv_cb().

Fix bug in _getgroup_done() callback function for multi-zb model.

Consolidate barrier, broadcast, and reduce into common _reduce().

Modify _reduce() to require that supplied *dataptr represent an array
with zb->simcount elements. There is no change for production
(simcount==1), but simulation tests must now provide this as an array
rather than a pointer to a value. This allows heterogeneous simulated
data values. Prior lack of this is a likely reason the zbsend_up() bug
was not caught.

Add simulated test data arrays with shuffling to ensure that correct
data is being broadcast (NETSIM).

Add simulated reduce test (NETSIM).

Add distributed reduce test (production model).

Signed-off-by: Joe Nemeth <joseph.nemeth@hpe.com>