prov/hook/dmabuf_peer_mem: Close dmabuf fd when no longer in use
zeMemGetIpcHandle() allocates a new dma-buf fd every time it is called.
It is important to close the fd when it is no longer in use, otherwise
the process may run out of file descriptors.

Add a mutex to protect the dmabuf registry access to prevent the fd from
being incorrectly closed in multi-threaded runs.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>