fabtests/shm_sighandler_test: fix initialization issues
The shm_sighandler_test was failing intermittently because the usleep time
to let the provider initialize was usually not enough time so the intialization
issues were avoided. In the case where shm was able to initialize, it would fail
because of a couple issues - INIT_OPTS was not set so it complained that we are
not using a rx cq; fi_recv was using 0 instead of remote_fi_addr to post the receive,
causing a segfault on address lookup because the address did not exist.

This patch increases the usleep to make sure shm always initializes fully.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>