core/pollfds: Fix signaling race
In ofi_pollfds_waitall, after the thread returns from poll it
first processes any queued work from the work_item_list.  It
then releases the pollfds lock prior to reseting the signal.
However, this leaves a gap where an item could be added to the
work list after processing occurs, and the resulting signal
could be cleared.  The result is that the thread could loop
back into the poll() call and block indefinitely (timeout = -1),
and the new work item will be left pending.

To fix, we reset the signal while still holding the pollfds
lock.  To be safer, we also swap the ordering.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>