core/pollfds: Adjust number of events to check
In the pollfds abstraction, the return value from poll
indicates the number of events that were returned.  In
some cases, the events are for fd's that we've already
removed from the pollfds.  As a result, we skip these
events, however, we don't adjust the number of events
reported by poll.  This can cause us to look for events
that aren't there.

For example, if poll() returns 1, the code may see that
the event belongs to fd 3, but that fd 3 is no longer part of
the poll set.  Instead of stopping at this point, the code
continues to loop, checking all entries in the poll set for
events.  (This is partly due to incrementing 'found' inside
an if statement that may be false).

Optimize the code to stop check for events once all known
events have been identified.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>