pollfds: Skip signaling fd unless waiting
The signaling fd is used to wake-up any thread blocked in
poll.  We can skip checking it if we're calling poll with
a timeout of 0.

This is a minor optimization, but also helps in a rare
case where we have 1 thread blocking in poll, and another
thread calling poll without blocking.  In the latter case,
we skip checking the signaling fd, which allows the blocked
thread to respond to it.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>