prov/shm: Add HMEM IPC Caching Mechanism.
Opening and closing the HMEM IPC handles are expensive operations. They
should not be performed on every data operation.

This patch adds a caching mechanism. On the very first IPC handle open
call, the address is cached and the mapped memory address is stored. In
subsequent calls the cache is queried and the mapped address is retrieved.

On open handle the hmem specific code can return the actual length and
base address, which is cached. It is often the case that a big chunk of
device memory is allocated and managed by the app. This method allows the
libfabric code to recognize the initial allocation and only register it
once. Any address within this allocation is a cache hit.

In a future patch the caching mechanism can be used with memory
registration APIs. This design allows for caching to be invisible to the
user, thereby making the API simpler to use.

This is the same caching mechanism used in UCX.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>