hmem/hip: Reference HIP implementation
ROCM can be supported in libfabric via the HSA or HIP APIs. Currently the
HSA API is preferred because it doesn't use the BLIT kernel by default,
but uses the SDMA engine. This is a desirable behavior, since the BLIT
kernel would use GPU resources that could otherwise be used for
application computational needs, rather than data movement. If all GPU
resources are busy and the BLIT kernel is used, then the transfer could be
delayed.

The BLIT kernel is able to make use of all the xGMI3 links between the
GPUs. There is an outstanding feature development on the AMD side to allow
the DMA engine to do the same. However, currently it doesn't and therefore
there is a limit on the performance using the DMA engine, approx ~50 GB/s.
HIP APIs can achieve approx 150 GB/s between GCDs on the same OAM, since
it can multi-rail across the 4 xGMI3 links. Each of which can provide up
to 50 GB/s.

It is conceivable to have both implementations in libfabric and allow the
app to determine which one to use based on its forcasted use of GPU
resources.

At this time however, this patch is submitted as a reference
implementation. A decision whether to include it or not, can be made at a
later date

Signed-off-by: Amir Shehata <shehataa@ornl.gov>