fabtests/multinode: Use FI_DELIVERY_COMPLETE for barrier
Use delivery complete semantics for the barrier implementation
to ensure that all sends have been received by the peer.

Reverse checking for the receive completion prior to checking
for sends.  This is needed to avoid a potential deadlock
assuming that sends are buffered prior to the receive
completions being processed.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>