prov/tcp: Verify ep state prior to progressing socket data
Skip progressing data transfers unless the ep is in the
connected state.  This avoids the potential of sending data
that may be interpreted as the start of a message by the peer,
and processing received data in the middle of the data stream
as the start of a new message.

There are no identified issues in the current code requiring
this change; however, problems persist with application segfaults
when connections are abruptly terminated.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>