prov/shm: Enable CMA even when HMEM is requested
Currently CMA is disabled if HMEM is requested. The reason for this is to
avoid inconsistent behavior if CMA is used when the receiving buffer is a
device buffer. The operation would fail in this case.

This patch removes this restriction.

TODO: A future patch will ensure the Host to Device scenario is handled
properly.

Signed-off-by: Amir Shehata <shehataa@ornl.gov>